fix wrong type in libf

+++ libft/includes/libft.h
@@ -40,9 +40,9 @@ int				ft_memcmp(const void *s1, const void *s2, size_t n);
 
 // STRINGS
 
-int				ft_strlen(const char *str);
+size_t			ft_strlen(const char *str);
 int				ft_strcmp(const char *s1, const char *s2);
-int				ft_strncmp(const char *s1, const char *s2, unsigned int n);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
 size_t			ft_strlcat(char *dest, const char *src, size_t size);
 char			*ft_strdup(const char *src);
